Here is the algorithm
This is indeed a very tricky problem. Generally if you follow a bruteforce approach than it will yield time limit exceeded error. I got stucked with this problem for a lot of days just.So here is the algorithm:
1); make the palindrome of the given no by diving the number into half and mirroring it. Here 2 cases will arise.
e.g1- if no is 4321 than palin is 4334, and that is the correct answer, however
e.g2- if no is 1234 than palin os 1221, but it not the desire result as it is less than the original number, so  here what we will do:
         we will take the center digit and increment if(ofcourse in case of even length 2 digits will increment as here it will become 1331 and that is the             answer and for odd no like 12345, it will become 12321 but that is less so we will increment centre digit as 12421 which is the answer)
e.g3- if the centre digit contains the 9, than do it as for example the no is 1999, after palin connversion it becomes 1991, than increment centre it becomes 1001 than iterate one more increment on left most bit and right most bit as 2002 which will be the answer
e.g4-if the no is 9999, than the answer will be 10001.